moved gate classes into own files
